requirements.txt
python-dotenv==1.0.0
psycopg2-binary==2.9.9
osmium==3.7.0
requests
psutil>=5.9.0

